Bookmarks almost ready, need add table in db.

+++ bot/bot_callbacks.py
@@ -1,19 +1,25 @@
-from telegram import InlineKeyboardMarkup
-
 from .keyboards import (
+    add_bookmarks_button,
     continents_buttons,
     countries_buttons,
     get_resort_info,
+    resort_to_bookmarks,
     resorts_buttons,
 )
 
 
+def available_commands(update, context):
+    answer = "/start - start to conversation\n/info - search resort by regions"
+    return update.message.reply_text(answer)
+
+
 def start(update, context):
-    user_says = " ".join(context.args)
-    update.message.reply_text("Hello! " + user_says + "!!!")
+    username = update.message.chat.first_name
+    update.message.reply_text(f"Hello {username}!")
     context.bot.send_message(
         chat_id=update.effective_chat.id, text="I'm a bot, please talk to me!"
     )
+    return available_commands(update, context)
 
 
 def choose_buttons(next_step):
@@ -27,21 +33,27 @@ def choose_buttons(next_step):
     raise ValueError("Unknown type of next step")
 
 
-def button(update, context):
+def manage_callback(update, context):
     query = update.callback_query
     query.answer()
     command = query.data
+    if command.split(sep=":")[0] == "add_bookmarks":
+        # 128609524
+        user_id = query.message.chat.id
+        resort_name = command.split(sep=":")[1]
+        return resort_to_bookmarks(user_id=user_id, resort_id=resort_name)
     if command == "cancel":
         return update.callback_query.delete_message()
     if command.split(sep=":")[0] == "resort":
-        resort_info = get_resort_info(command.split(sep=":")[1])
-        return update.callback_query.message.reply_text(resort_info)
-    keyboard = choose_buttons(command)
-    reply_markup = InlineKeyboardMarkup(keyboard)
+        resort_name = command.split(sep=":")[1]
+        resort_info = get_resort_info(resort_name)
+        return update.callback_query.message.reply_text(
+            resort_info, reply_markup=add_bookmarks_button(resort_name)
+        )
+    reply_markup = choose_buttons(command)
     return query.edit_message_reply_markup(reply_markup=reply_markup)
 
 
 def select_continents(update, context):
-    keyboard = continents_buttons()
-    reply_markup = InlineKeyboardMarkup(keyboard)
+    reply_markup = continents_buttons()
     update.message.reply_text("Please choose:", reply_markup=reply_markup)
